refactor: header component
